requirements.txt
pyserial==3.0.1
six==1.10.0
